img2Objtext/requirements.txt
opencv-python
numpy
torch
pillow
pandas
requests
psutil
torchvision
PyYAML
tqdm
ultralytics
